pyproject.toml
[project]
name = "indappproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "freesimplegui>=5.2.0.post1",
    "pillow>=11.3.0",
    "staticmap>=0.5.7",
]
